% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_plot_md.r
\name{display_plot_md}
\alias{display_plot_md}
\title{Cat path to a plot for Markdown doc}
\usage{
display_plot_md(x)
}
\arguments{
\item{x}{character.
Path to a plot.}
}
\description{
Useful so you don't need to
remember the syntax.
}
\examples{
display_plot_md(here::here("test.png"))
}
