% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_file.r
\name{get_relative_path}
\alias{get_relative_path}
\title{Get relative path}
\usage{
get_relative_path(path, path_base)
}
\arguments{
\item{path}{character.
Path to object for which we require a relative path.}

\item{path_base}{character.
Path against which the relative path should be calculated.}
}
\description{
Get the path to a file/directory
relative to some sub-path.
}
