% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.r
\name{save_plot}
\alias{save_plot}
\title{Save plot in multiple formats using cowplot::save_plot}
\usage{
save_plot(filename, plot, device = c("png", "pdf"), ...)
}
\arguments{
\item{filename}{character. Name of plot (including directory, if
not to save in working directory).}

\item{plot}{gg object. Output from either ggplot2::ggplot
or cowplot::plot_grid.}

\item{device}{character vector. Devices to save in (i.e. plot format).
Default is \code{c("png", "pdf")}.}

\item{...}{passed on to \code{cowplot::save_plot}.}
}
\value{
\code{invisible(filename)}.
}
\description{
Wrapper around \code{cowplot::save_plot} to save plots
in multiple formats.
}
