% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_output_dir.r
\name{setup_bookdown_output_dir}
\alias{setup_bookdown_output_dir}
\alias{set_up_bookdown_output_dir}
\title{Set up bookdown output directories}
\usage{
setup_bookdown_output_dir(
  sub = c(dir_m_fig = "manu/fig", dir_m_tbl = "manu/tbl", dir_mn_fig = "manu_n/fig",
    dir_mn_tbl = "manu_n/tbl")
)
}
\arguments{
\item{sub}{named character vector.
The names are the names of the objects to be created, and
the values are paths relative to the report output directory.}
}
\value{
\code{invisible(TRUE)}. Side effects are
the creation of directories and saving
of objects specifying those directories
to global environment.
}
\description{
Create objects that
give paths relative to the bookdown output directory.
In addition, it creates those directories if they
do not already exist.

Useful to save results to such folders
so that the resultant bookdown output directory
is portable, i.e. can be moved around and the
path references to the figures and tables
still work.

If the output directory is unspecified (i.e. there is
no entry \code{output_dir} in \code{_bookdown.yml}),
then the output directory is taken to be the default \code{_book}.
}
\examples{
wd_temp <- file.path(tempdir(), "test_setup_output_dir")
if (!dir.exists(wd_temp)) {
  dir.create(wd_temp, recursive = TRUE)
}
file.create("_bookdown.yml")
setup_bookdown_output_dir()
dir_m_fig

}
