% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_output_dir.r
\name{setup_output_dir}
\alias{setup_output_dir}
\alias{set_up_output_dir}
\title{Set up output directories}
\usage{
setup_output_dir(dir_base = here::here("_book"), prefix = ".dir_bd_")

set_up_output_dir(dir_base = here::here("_book"), prefix = ".dir_bd_")
}
\arguments{
\item{dir_base}{character. Path to base directory where objects
are to be saved.
Default is \code{here::here("_book")}.}

\item{prefix}{character. Prefix for naming
objects specifying paths to particular directories.
Default is \code{.dir_bd}. This results in objects
\code{.dir_bd_manu}, \code{.dir_bd_manu_n},
\code{.dir_bd_manu_fig}, \code{dir_bd_manu_tbl},
\code{.dir_bd_manu_n_fig} and \code{dir_bd_manu_n_tbl},
which provide the absolute paths.
Corresponding objects ending in "_rel" (e.g.
\code{.dir_bd_manu_rel}) specify the
relative paths (relative to \code{here::here()}).}
}
\value{
\code{invisible(TRUE)}. Side effects are
the creation of directories and saving
of objects specifying those directories
to global environment.
}
\description{
Set up \code{bookdown} manuscript
and non-manuscript figure and table output directories.
Creates directories, and saves objects specifying these directories
into global environment. See \code{prefix} for details
of objects saved.
}
